change chosen size[C

+++ pyprob/nn/dataset.py
@@ -327,7 +327,8 @@ class DistributedTraceBatchSampler(Sampler):
         # List of all minibatches in the whole dataset, where each minibatch is a list of trace indices
         #self._batches = list(util.chunks(offline_dataset._sorted_indices, batch_size))
         batches_skip = (len(offline_dataset._sorted_indices)/batch_size)%self._world_size
-        chosen_size = len(offline_dataset._sorted_indices)-batches_skip * batch_size
+        #chosen_size = len(offline_dataset._sorted_indices)-batches_skip * batch_size
+        chosen_size = int(len(offline_dataset._sorted_indices)-batches_skip * batch_size)
         chosen_indices = random.sample(list(offline_dataset._sorted_indices), chosen_size)
         self._batches = list(util.chunks(chosen_indices, batch_size))
         # Discard last minibatch if it's smaller than batch_size
